17-02-2024 Add 'config' and improved 'entity', 'repository'.

+++ src/main/java/com/techmatrix18/service/InterfaceUserService.java
@@ -5,9 +5,9 @@ import java.util.List;
 
 public interface InterfaceUserService {
     List<User> getAllUsers();
-    User getUserById(long userId);
+    User getUserById(Long userId);
     boolean addUser(User user);
     void updateUser(User user);
-    void deleteUser(int userId);
+    void deleteUser(Long userId);
 }
